Adapt path string to pass it to Python 3's C API
Since the function PySys_SetPath() in Python 3 only accepts a pointer
to a wide character array, we need to convert a narrow string to a
wide one. Although this conversion is not straightforward, components
defined in <codecvt> header are used as shown in, for example,
https://stackoverflow.com/a/18597384/14398042.

Note that the header <codecvt> is available since C++11 and is
deprecated in C++17.