Fix to check out specific tag instead of branch
So far, to clone a specific version of Choreonoid the corresponding
branch is checked out. But branches for version 1.5 and 1.6 are only
available in s-nakaoka/choreonoid, which requires us to indicate
s-nakaoka/choreonoid or choreonoid/choreonoid to be checked out in
accordance with the current job. This commit will fix the issue by
specifying a tag in choreonoid/choreonoid instead of a branch.