Break down matrix strategy jobs into runner OSs
Since we need to create different environments to run jobs on CI
runner OSs, many if-then conditional rules are required when keeping
all jobs unified in one job, which leads complexity and less
maintainability. So separate the CI jobs defined as matrix strategy
into jobs based on the runner OS.

Also, this commit adds `workflow_dispatch` trigger which enables us to
run the workflow manually and prevents from running it when pushing on
brances except `master` and `*-ci`.